Manual timestamp entry for a11y or power users.
Also factors out the ElapsedTimeUtils class, which
could later be replaced by Joda-Time, and writes
some tests for that class.

BUG=32648113

+++ OpenScienceJournal/whistlepunk_library/src/sharedTest/java/com/google/android/apps/forscience/whistlepunk/sensorapi/StubInteractionListener.java
@@ -35,7 +35,7 @@ class StubInteractionListener implements ExternalAxisController.InteractionListe
     }
 
     @Override
-    public void onZoom(double xMin, double xMax) {
+    public void onZoom(long xMin, long xMax) {
 
     }
